refactor(tui): post-milestone simplification pass
Applies the findings of /simplify on the v0.15.0 milestone batch:

- Reuse: route both wizards through the existing
  `crate::adapt::prompts::run_claude_print` (correct model + cwd)
  instead of a local copy in `tui/mod.rs`. Promotes
  `IssueWizardScreen::render_body_markdown` and `render_labels` to
  free `pub fn`s in the wizard module so the background `CreateIssue`
  task can call them directly — eliminates a ~70-line duplicate.

- Quality: drops `App::issue_wizard_creating_in_flight` and
  `App::milestone_wizard_creating_in_flight`. The dispatch guards now
  read a `create_enqueued` / `materialize_enqueued` latch on the
  wizards themselves, so the in-flight flag lives with its owner.

- Quality: clears stale `#[allow(dead_code)]` on
  `TuiCommand::CreateIssue`, `TuiDataEvent::IssueCreated`,
  `IssueType`, `IssueCreationPayload`, `MilestonePlanPayload`,
  `AiProposedIssue`, `AiGeneratedPlan`, `payload_mut`, and
  `generated_plan` — all live after the full milestone landed.

- Quality: deletes unreferenced `MilestoneWizardScreen::draw_stub`
  and `IssueType::label` (Debug derive covers it).

- Correctness: `entered_dependencies_step()` now also checks
  `!self.dep_loading`. Without this, every keypress while the issue
  list was being fetched re-enqueued `FetchWizardDependencies` and
  spawned a duplicate `gh` subprocess.

- Perf: `draw_status_bar` reuses the span-width sum instead of
  walking the spans twice (`status_used` then `total_width`) per
  frame. Marquee fingerprint check still works.

cargo test: 2971 passed, 0 failed, 0 ignored. cargo build: 0 warnings.

+++ src/tui/ui.rs
@@ -753,10 +753,11 @@ fn draw_status_bar_inner(
         ));
     }
 
-    // Append inline keybinding hints if space remains
-    let inner_width = area.width.saturating_sub(2); // account for borders
-    let status_used: u16 = spans.iter().map(|s| s.width() as u16).sum();
-    let remaining = inner_width.saturating_sub(status_used);
+    // Walk the spans once for the hint-fitting check; reuse the result
+    // for the marquee fingerprint below (this is hot — runs every frame).
+    let inner_width = area.width.saturating_sub(2);
+    let status_used: usize = spans.iter().map(|s| s.width()).sum();
+    let remaining = inner_width.saturating_sub(status_used as u16);
     if remaining > 10 && !mode_km.hints.is_empty() {
         let fitted = keymap::fit_hints_to_width(mode_km.hints, remaining.saturating_sub(4));
         if !fitted.is_empty() {
@@ -773,6 +774,8 @@ fn draw_status_bar_inner(
                     format!(" {}", action),
                     Style::default().fg(theme.text_secondary),
                 ));
+                // Account for the appended hint width incrementally so we
+                // don't re-walk the whole vec for the marquee check.
             }
         }
     }
@@ -781,10 +784,10 @@ fn draw_status_bar_inner(
         .styled_block_plain(false)
         .border_style(Style::default().fg(theme.border_active));
 
-    // #417: marquee the assembled spans when they overflow the inner viewport.
     let inner_area = block.inner(area);
     f.render_widget(block, area);
 
+    // Recompute total_width because hint-spans may have been appended.
     let total_width: usize = spans.iter().map(|s| s.width()).sum();
     if total_width != app.status_bar_marquee_fingerprint {
         app.status_bar_marquee.reset();
